Replier function is now part of universal bot tools. Refactored function names to make them easier to understand.
