+++ src/botTools/memory/curated-data.js
