--- src/core/index.js
@@ -1,6 +1,0 @@
-/**
- * Export every route manager
- */
-module.exports = {
-  dbApi: require('./core-api'),
-};
