+++ src/rooCoreApi/index.js
@@ -1,0 +1,4 @@
+/**
+ * Export every route manager
+ */
+module.exports.rooCoreApi = require('./api');
